device: Move downlink mac handler from Configuration to mac
Downlink handling should be done at least at mac scope.
